Integrates `smartyresponseprocessing` Into main
This change folds `smartyresponseprocessing` package directly into the `smarty-usps-comparison/main.go` file in order to consolidate all tool-specific logic into one main script.

Tested: Executed the script to confirm that the integrated processing logic functions correctly.
