+++ cmd/smarty-usps-comparison/main.go
@@ -10,11 +10,40 @@ import (
 	"log"
 	"net/http"
 	"os"
+	"sort"
 	"strings"
 	"time"
 
-	"github.com/corbaltcode/usps/internal/smartyresponseprocessing"
+	"golang.org/x/exp/slices"
 )
+
+type SmartyResponse struct {
+	Zipcodes []struct {
+		Zipcode           string  `json:"zipcode"`
+		ZipcodeType       string  `json:"zipcode_type"`
+		DefaultCity       string  `json:"default_city"`
+		CountyFIPS        string  `json:"county_fips"`
+		CountyName        string  `json:"county_name"`
+		StateAbbreviation string  `json:"state_abbreviation"`
+		State             string  `json:"state"`
+		Latitude          float64 `json:"latitude"`
+		Longitude         float64 `json:"longitude"`
+		Precision         string  `json:"precision"`
+		AlternateCounties []struct {
+			CountyFIPS        string `json:"county_fips"`
+			CountyName        string `json:"county_name"`
+			StateAbbreviation string `json:"state_abbreviation"`
+			State             string `json:"state"`
+		} `json:"alternate_counties"`
+	} `json:"zipcodes"`
+}
+
+type ZipcodeResult struct {
+	Zipcode     string
+	USPSFips    []string
+	SmartyFips  []string
+	HasMismatch bool
+}
 
 type ZipcodeRequest struct {
 	Zipcode string `json:"zipcode"`
@@ -34,7 +63,7 @@ func NewSmartyClient(authId, authToken string) *SmartyClient {
 	}
 }
 
-func (client *SmartyClient) QueryBatch(zips []string) ([]smartyresponseprocessing.SmartyResponse, error) {
+func (client *SmartyClient) QueryBatch(zips []string) ([]SmartyResponse, error) {
 	var payload []ZipcodeRequest
 	for _, zip := range zips {
 		payload = append(payload, ZipcodeRequest{Zipcode: zip})
@@ -45,6 +74,7 @@ func (client *SmartyClient) QueryBatch(zips []string) ([]smartyresponseprocessin
 	}
 
 	apiURL := fmt.Sprintf("%s?auth-id=%s&auth-token=%s", client.BaseURL, client.AuthId, client.AuthToken)
+
 	req, err := http.NewRequest("POST", apiURL, bytes.NewBuffer(jsonPayload))
 	if err != nil {
 		return nil, fmt.Errorf("failed to create HTTP request: %w", err)
@@ -70,7 +100,7 @@ func (client *SmartyClient) QueryBatch(zips []string) ([]smartyresponseprocessin
 		return nil, fmt.Errorf("API request failed with status %d: %s", resp.StatusCode, body)
 	}
 
-	var smartyResponses []smartyresponseprocessing.SmartyResponse
+	var smartyResponses []SmartyResponse
 	err = json.Unmarshal(body, &smartyResponses)
 	if err != nil {
 		return nil, fmt.Errorf("failed to unmarshal response: %w", err)
@@ -120,6 +150,36 @@ func parseCSVList(dataString string) []string {
 	return items
 }
 
+func ProcessSmartyResponse(responseBody []SmartyResponse, zipToCounties map[string][]string, yield func(ZipcodeResult)) error {
+	for _, response := range responseBody {
+		for _, zipcode := range response.Zipcodes {
+			result := ZipcodeResult{
+				Zipcode:  zipcode.Zipcode,
+				USPSFips: zipToCounties[zipcode.Zipcode],
+			}
+
+			// Extract the last three digits of the Smarty Fips code, which represent the county.
+			smartyFIPS := []string{zipcode.CountyFIPS[len(zipcode.CountyFIPS)-3:]}
+			for _, altCounty := range zipcode.AlternateCounties {
+				smartyFIPS = append(smartyFIPS, altCounty.CountyFIPS[len(altCounty.CountyFIPS)-3:])
+			}
+			result.SmartyFips = smartyFIPS
+
+			result.HasMismatch = slicesDiffer(result.USPSFips, result.SmartyFips)
+
+			yield(result)
+		}
+	}
+
+	return nil
+}
+
+func slicesDiffer(uspsFIPS, smartyFIPS []string) bool {
+	sort.Strings(uspsFIPS)
+	sort.Strings(smartyFIPS)
+	return !slices.Equal(uspsFIPS, smartyFIPS)
+}
+
 func mustGetenv(key string) string {
 	v, ok := os.LookupEnv(key)
 	if !ok {
@@ -154,7 +214,7 @@ func main() {
 		log.Fatalf("Error reading CSV file: %v", err)
 	}
 
-	yield := func(result smartyresponseprocessing.ZipcodeResult) {
+	yield := func(result ZipcodeResult) {
 		if result.HasMismatch {
 			log.Printf("%+v", result)
 		}
@@ -175,7 +235,7 @@ func main() {
 			log.Printf("Error querying Smarty API: %v", err)
 		}
 
-		smartyresponseprocessing.ProcessSmartyResponse(responseBody, zipToCounty, yield)
+		ProcessSmartyResponse(responseBody, zipToCounty, yield)
 
 		log.Printf("%v zips processed.", i+batchSize)
